feat(Site) - #505 - fixed mailing list to make it a reusable backend function and to be server side so nothing's ever exposed
