+++ functions/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/bootstrapadmin.ts","./src/config.ts","./src/index.ts","./src/_shared/aichat.ts","./src/_shared/aiimageanalysis.ts","./src/_shared/belrosefields.ts","./src/_shared/blockchainaddresses.ts","./src/_shared/converttofhir.ts","./src/_shared/credibility.ts","./src/_shared/deletion.ts","./src/_shared/index.ts","./src/_shared/notifications.ts","./src/_shared/permissions.ts","./src/_shared/recordrequest.ts","./src/_shared/subject.ts","./src/_shared/timestamp.ts","./src/handlers/aichat.ts","./src/handlers/belrosefields.ts","./src/handlers/checkemailregistrationstatus.ts","./src/handlers/converttofhir.ts","./src/handlers/createguestinvite.ts","./src/handlers/createrecordrequest.ts","./src/handlers/guestpasswordupdate.ts","./src/handlers/healthcheck.ts","./src/handlers/image.ts","./src/handlers/memberregistry.ts","./src/handlers/paymaster.ts","./src/handlers/redeemguestinvite.ts","./src/handlers/refinerecord.ts","./src/handlers/sendalphaapprovalemail.ts","./src/handlers/sendmailinglistconfirmationemail.ts","./src/handlers/sendpasswordchangeemail.ts","./src/handlers/sendshareinvitationemail.ts","./src/handlers/sendwaitlistconfirmationemail.ts","./src/handlers/setadminclaim.ts","./src/handlers/stripeidentity.ts","./src/handlers/verification.ts","./src/handlers/wallet.ts","./src/notifications/emailutils.ts","./src/notifications/notificationutils.ts","./src/notifications/emails/credibilityemailtemplates.ts","./src/notifications/emails/permissionemailtemplate.ts","./src/notifications/emails/recorddeletionemailtemplates.ts","./src/notifications/emails/recordeditemailtemplate.ts","./src/notifications/emails/recordrequestemailtemplate.ts","./src/notifications/emails/subjectemailtemplates.ts","./src/notifications/emails/trusteeemailtemplates.ts","./src/notifications/triggers/credibilitynotificationtrigger.ts","./src/notifications/triggers/deleterecordnotificationtrigger.ts","./src/notifications/triggers/permissionnotificationtrigger.ts","./src/notifications/triggers/recordeditnotificationtrigger.ts","./src/notifications/triggers/requestrecordnotificationtrigger.ts","./src/notifications/triggers/subjectnotificationtrigger.ts","./src/notifications/triggers/trusteenotificationtriggers.ts","./src/services/anthropicservice.ts","./src/services/backendwalletservice.ts","./src/services/claudechatservice.ts","./src/services/geminiservice.ts","./src/services/openaiservice.ts","./src/utils/aichatprompt.ts","./src/utils/guestaccountutils.ts","./src/utils/recordaiprompts.ts"],"version":"5.9.3"}
+{"root":["./src/bootstrapadmin.ts","./src/config.ts","./src/index.ts","./src/_shared/aichat.ts","./src/_shared/aiimageanalysis.ts","./src/_shared/belrosefields.ts","./src/_shared/blockchainaddresses.ts","./src/_shared/converttofhir.ts","./src/_shared/credibility.ts","./src/_shared/deletion.ts","./src/_shared/index.ts","./src/_shared/notifications.ts","./src/_shared/permissions.ts","./src/_shared/recordrequest.ts","./src/_shared/subject.ts","./src/_shared/timestamp.ts","./src/handlers/addtomailinglist.ts","./src/handlers/aichat.ts","./src/handlers/belrosefields.ts","./src/handlers/checkemailregistrationstatus.ts","./src/handlers/converttofhir.ts","./src/handlers/createguestinvite.ts","./src/handlers/createrecordrequest.ts","./src/handlers/guestpasswordupdate.ts","./src/handlers/healthcheck.ts","./src/handlers/image.ts","./src/handlers/memberregistry.ts","./src/handlers/paymaster.ts","./src/handlers/redeemguestinvite.ts","./src/handlers/refinerecord.ts","./src/handlers/sendalphaapprovalemail.ts","./src/handlers/sendpasswordchangeemail.ts","./src/handlers/sendshareinvitationemail.ts","./src/handlers/sendwaitlistconfirmationemail.ts","./src/handlers/setadminclaim.ts","./src/handlers/stripeidentity.ts","./src/handlers/verification.ts","./src/handlers/wallet.ts","./src/notifications/emailutils.ts","./src/notifications/notificationutils.ts","./src/notifications/emails/credibilityemailtemplates.ts","./src/notifications/emails/permissionemailtemplate.ts","./src/notifications/emails/recorddeletionemailtemplates.ts","./src/notifications/emails/recordeditemailtemplate.ts","./src/notifications/emails/recordrequestemailtemplate.ts","./src/notifications/emails/subjectemailtemplates.ts","./src/notifications/emails/trusteeemailtemplates.ts","./src/notifications/triggers/credibilitynotificationtrigger.ts","./src/notifications/triggers/deleterecordnotificationtrigger.ts","./src/notifications/triggers/permissionnotificationtrigger.ts","./src/notifications/triggers/recordeditnotificationtrigger.ts","./src/notifications/triggers/requestrecordnotificationtrigger.ts","./src/notifications/triggers/subjectnotificationtrigger.ts","./src/notifications/triggers/trusteenotificationtriggers.ts","./src/services/anthropicservice.ts","./src/services/backendwalletservice.ts","./src/services/claudechatservice.ts","./src/services/geminiservice.ts","./src/services/openaiservice.ts","./src/utils/aichatprompt.ts","./src/utils/guestaccountutils.ts","./src/utils/recordaiprompts.ts"],"version":"5.9.3"}
